changes needed in tests for Visual Studio 2013

+++ build.vc12/flint_config/flint_config.py
@@ -21,9 +21,9 @@ from time import sleep
 # for script debugging
 debug = False
 # what to build
-build_lib = True
-build_dll = True
-build_tests = False
+build_lib = False
+build_dll = False
+build_tests = True
 build_profiles = False
 
 # add user choice
@@ -40,6 +40,10 @@ try:
   input = raw_input
 except NameError:
   pass
+
+app_type, lib_type, dll_type = 0, 1, 2
+app_str = ('Application', 'StaticLibrary', 'DynamicLibrary')
+app_ext = ('.exe', '.lib', '.dll')
 
 # copy from file ipath to file opath but avoid copying if
 # opath exists and is the same as ipath (this is to avoid
@@ -108,8 +112,6 @@ def filter_folders(cf_list, outf):
       if d:
         outf.write(f2.format(d))
     outf.write(f3)
-
-filter_hdr_item = r'    <ClInclude Include="..\..\{}">\n      <Filter>Header Files</Filter>\n    </ClInclude>\n'
 
 def filter_headers(hdr_list, relp, outf):
 
@@ -206,16 +208,16 @@ def vcx_default_cpp_props(outf):
 '''
   outf.write(f1)
 
-def vcx_library_type(plat, is_dll, outf):
+def vcx_library_type(plat, proj_type, outf):
 
   f1 = r'''  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='{1:s}|{0:s}'" Label="Configuration">
-    <ConfigurationType>{2:s}Library</ConfigurationType>
+    <ConfigurationType>{2:s}</ConfigurationType>
     <PlatformToolset>v120</PlatformToolset>
     </PropertyGroup>
 '''
   for pl in plat:
     for conf in ('Release', 'Debug'):
-      outf.write(f1.format(pl, conf, 'Dynamic' if is_dll else 'Static'))
+      outf.write(f1.format(pl, conf, app_str[proj_type]))
 
 def vcx_cpp_props(outf):
 
@@ -233,7 +235,7 @@ def vcx_user_props(plat, outf):
     for conf in ('Release', 'Debug'):
       outf.write(f1.format(pl, conf))
 
-def vcx_target_name_and_dirs(name, plat, is_dll, outf):
+def vcx_target_name_and_dirs(name, plat, proj_type, outf):
 
   f1 = r'''  <PropertyGroup>
     <_ProjectFileVersion>10.0.21006.1</_ProjectFileVersion>
@@ -247,10 +249,10 @@ def vcx_target_name_and_dirs(name, plat, is_dll, outf):
   outf.write(f1)
   for pl in plat:
     for conf in ('Release', 'Debug'):
-      outf.write(f2.format(pl, conf, name, ('dll' if is_dll else 'lib')))
+      outf.write(f2.format(pl, conf, name, app_ext[proj_type]))
   outf.write(f3)
 
-def compiler_options(plat, is_dll, is_debug, inc_dirs, outf):
+def compiler_options(plat, proj_type, is_debug, inc_dirs, outf):
 
   f1 = r'''    <ClCompile>
     <Optimization>{0:s}</Optimization>
@@ -263,12 +265,17 @@ def compiler_options(plat, is_dll, is_debug, inc_dirs, outf):
     </ClCompile>
 '''
 
-  if is_dll:
-    s1 = 'DEBUG;HAVE_CONFIG_H;MSC_BUILD_DLL;'
-    s2 = 'DLL'
-  else:
-    s1 = 'DEBUG;_LIB;HAVE_CONFIG_H;'
+  if proj_type == app_type:
+    s1 = 'DEBUG;WIN32;_CONSOLE'
     s2 = ''
+  if proj_type == dll_type:
+    s1 = 'DEBUG;WIN32;HAVE_CONFIG_H;MSC_BUILD_DLL;'
+    s2 = 'DLL'
+  elif proj_type == lib_type:
+    s1 = 'DEBUG;WIN32;_LIB;HAVE_CONFIG_H;'
+    s2 = ''
+  else:
+    pass
   if plat == 'x64':
     s1 = s1 + '_WIN64;'
   if is_debug:
@@ -298,7 +305,7 @@ def vcx_post_build(outf):
 
   outf.write(f1)
 
-def vcx_tool_options(plat, is_dll, postbuild, inc_dirs, link_libs, outf):
+def vcx_tool_options(plat, proj_type, postbuild, inc_dirs, link_libs, outf):
 
   f1 = r'''  <ItemDefinitionGroup Condition="'$(Configuration)|$(Platform)'=='{1:s}|{0:s}'">
 '''
@@ -307,8 +314,8 @@ def vcx_tool_options(plat, is_dll, postbuild, inc_dirs, link_libs, outf):
   for pl in plat:
     for is_debug in (False, True):
       outf.write(f1.format(pl, 'Debug' if is_debug else 'Release'))
-      compiler_options(pl, is_dll, is_debug, inc_dirs, outf)
-      if is_dll:
+      compiler_options(pl, proj_type, is_debug, inc_dirs, outf)
+      if proj_type != lib_type:
         linker_options(link_libs, outf)
       if postbuild:
         vcx_post_build(outf)
@@ -364,7 +371,7 @@ def vcx_c_items(cf_list, plat, relp, outf):
       outf.write(f6)
   outf.write(f7)
 
-def gen_vcxproj(proj_name, file_name, guid, plat, is_dll, postbuild, hf_list, cf_list, inc_dirs, link_libs):
+def gen_vcxproj(proj_name, file_name, guid, plat, proj_type, postbuild, hf_list, cf_list, inc_dirs, link_libs):
 
   f1 = r'''<?xml version="1.0" encoding="utf-8"?>
 <Project DefaultTargets="Build" ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
@@ -382,18 +389,21 @@ def gen_vcxproj(proj_name, file_name, guid, plat, is_dll, postbuild, hf_list, cf
     vcx_proj_cfg(plat, outf)
     vcx_globals(proj_name, guid, outf)
     vcx_default_cpp_props(outf)
-    vcx_library_type(plat, is_dll, outf)
+    vcx_library_type(plat, proj_type, outf)
     vcx_cpp_props(outf)
     vcx_user_props(plat, outf)
     outf.write(f2)
-    vcx_target_name_and_dirs(proj_name, plat, is_dll, outf)
-    vcx_tool_options(plat, is_dll, postbuild, inc_dirs, link_libs, outf)
+    vcx_target_name_and_dirs(proj_name, plat, proj_type, outf)
+    vcx_tool_options(plat, proj_type, postbuild, inc_dirs, link_libs, outf)
     if hf_list:
       vcx_hdr_items(hf_list, relp, outf)
     vcx_c_items(cf_list, plat, relp, outf)
     outf.write(f3)
 
 # add project files to the solution
+
+folder_guid = "{2150E333-8FDC-42A3-9474-1A3956D46DE8}"
+vcxproject_guid = "{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}"
 
 s_guid = r'\s*(\{\w{8}-\w{4}-\w{4}-\w{4}-\w{12}\})\s*'
 s_name = r'\s*\"([a-zA-Z][-.\\_a-zA-Z0-9]*\s*)\"\s*'
@@ -405,17 +415,18 @@ re_fmap = compile(r'\s*' + s_guid + r'\s*=\s*' + s_guid)
 def read_solution_file(soln_name):
   fd, pd, p2f = {}, {}, {}
   solution_path = join(solution_dir, soln_name)
-  lines = open(solution_path).readlines()
-  for i, ln in enumerate(lines):
-    m = re_proj.search(ln)
-    if m: 
-      if m.group(2) == m.group(3):
-        fd[m.group(2)] = (m.group(1), m.group(4))
-      elif m.group(3).endswith('.vcxproj') or m.group(3).endswith('.pyproj'): 
-        pd[m.group(2)] = (m.group(1), m.group(3), m.group(4))
-    m = re_fmap.search(ln)
-    if m:
-      p2f[m.group(1)] = m.group(2)
+  if exists(solution_path):
+    lines = open(solution_path).readlines()
+    for i, ln in enumerate(lines):
+      m = re_proj.search(ln)
+      if m: 
+        if m.group(1) == folder_guid and m.group(2) == m.group(3):
+          fd[m.group(2)] = m.group(4)
+        elif m.group(3).endswith('.vcxproj') or m.group(3).endswith('.pyproj'): 
+          pd[m.group(2)] = (m.group(1), m.group(3), m.group(4))
+      m = re_fmap.search(ln)
+      if m:
+        p2f[m.group(1)] = m.group(2)
   return fd, pd, p2f
 
 sol_1 = '''Microsoft Visual Studio Solution File, Format Version 12.00
@@ -451,8 +462,8 @@ EndGlobal
 def write_solution_file(file_name, fd, pd, p2f):
   with open(join(solution_dir, file_name), 'w') as outf:
     outf.write(sol_1)
-    for f, (g1, g2) in fd.items():
-      outf.write(sol_2.format(g1, f, f, g2))
+    for f, g in fd.items():
+      outf.write(sol_2.format(folder_guid, f, f, g))
     for f, (g1, pn, g2) in pd.items():
       outf.write(sol_2.format(g1, f, pn, g2))
     outf.write(sol_3)
@@ -464,13 +475,13 @@ def add_proj_to_sln(soln_name, soln_folder, proj_name, file_name, guid):
   fd, pd, p2f = read_solution_file(soln_name)
   if soln_folder:
     if soln_folder in fd:
-      guid1, guid2 = fd[soln_folder]
+      f_guid = fd[soln_folder]
     else:
-      guid1, guid2 = '{' + str(uuid4()).upper() + '}', '{' + str(uuid4()).upper() + '}'
-      fd[soln_folder] = (guid1, guid2)
-  pd[proj_name] = ('{' + str(uuid4()).upper() + '}', file_name, guid)
+      f_guid = '{' + str(uuid4()).upper() + '}'
+      fd[soln_folder] = f_guid
+  pd[proj_name] = (vcxproject_guid, file_name, guid)
   if soln_folder:
-    p2f[guid] = guid2
+    p2f[guid] = f_guid
      
   write_solution_file(soln_name, fd, pd, p2f)
 
@@ -492,7 +503,7 @@ if build_lib:
   mode = ('win32', 'x64')
   inc_dirs = r'.\;..\;..\..\;..\..\..\mpir\lib\$(IntDir);..\..\..\mpfr\lib\$(IntDir);..\..\..\pthreads'
   link_libs = r'..\..\..\mpir\lib\$(IntDir)mpir.lib;..\..\..\mpfr\lib\$(IntDir)mpfr.lib;..\..\..\lib\pthreads.lib'
-  gen_vcxproj(proj_name, vcx_path, guid, mode, False, True, h, c, inc_dirs, link_libs)
+  gen_vcxproj(proj_name, vcx_path, guid, mode, lib_type, True, h, c, inc_dirs, link_libs)
   add_proj_to_sln(sln_name, '', proj_name, vcx_path, guid)
 
 if build_dll:
@@ -504,7 +515,7 @@ if build_dll:
   mode = ('win32', 'x64')
   inc_dirs = r'.\;..\;..\..\;..\..\..\mpir\dll\$(IntDir);..\..\..\mpfr\dll\$(IntDir);..\..\..\pthreads;'
   link_libs = r'..\..\..\mpir\dll\$(IntDir)mpir.lib;..\..\..\mpfr\dll\$(IntDir)mpfr.lib;..\..\..\lib\pthreads.lib;'
-  gen_vcxproj(proj_name, vcx_path, guid, mode, True, True, h, c, inc_dirs, link_libs)
+  gen_vcxproj(proj_name, vcx_path, guid, mode, dll_type, True, h, c, inc_dirs, link_libs)
   add_proj_to_sln(sln_name, '', proj_name, vcx_path, guid)
 
 def gen_test(sln_name, test_name, c_file):
@@ -516,7 +527,7 @@ def gen_test(sln_name, test_name, c_file):
   mode = ('win32', 'x64')
   inc_dirs = r'.\;..\..\;..\..\..\;..\..\..\..\mpir\lib\$(IntDir);..\..\..\..\mpfr\lib\$(IntDir);..\..\..\..\pthreads;'
   link_libs = r'..\..\..\..\mpir\lib\$(IntDir)mpir.lib;..\..\..\..\mpfr\lib\$(IntDir)mpfr.lib;..\..\..\..\lib\pthreads.lib;'
-  gen_vcxproj(proj_name, vcx_path, guid, mode, False, False, [], [('', c_file)], inc_dirs, link_libs)
+  gen_vcxproj(proj_name, vcx_path, guid, mode, app_type, False, [], [('', c_file)], inc_dirs, link_libs)
   return vcx_path, guid
 
 #   ('', 'test\\t-udiv_qrnnd_preinv.c')
@@ -530,24 +541,24 @@ def gen_tests(sln_name, c_files):
   for name, fpath in c_files:
     if name:
       if name in fd:
-        guid1, guid2 = fd[name]
+        f_guid = fd[name]
       else:
-        guid1, guid2 = '{' + str(uuid4()).upper() + '}', '{' + str(uuid4()).upper() + '}'
-        fd[name] = (guid1, guid2)
+        f_guid = '{' + str(uuid4()).upper() + '}'
+        fd[name] = f_guid
     _, t = split(fpath)
     if t[:2] not in ('t-', 'p-'):
       continue
     t = t[2:].replace('.c', '')
     p_name = (name + '_' + t) if name else t
     vcx_name, guid = gen_test(sln_name, p_name, fpath)
-    pd[p_name] = ('{' + str(uuid4()).upper() + '}', vcx_name, guid)
+    pd[p_name] = (vcxproject_guid, vcx_name, guid)
     if name:
-      p2f[guid] = guid2
+      p2f[guid] = f_guid
 
   write_solution_file(sln_name, fd, pd, p2f)
 
 if build_tests:
-  gen_tests('flint-tests2.sln', t[:10])
+  gen_tests('flint-tests.sln', tx)
 
 if debug:
   print('\nC++ files')
